successfully sent email, changed view in frontend to accept inputs if have shared itineraries

+++ frontend/src/pages/MyTripsPage.js
@@ -59,11 +59,15 @@ function MyTripsPage() {
         setLoading(false);
       })
       .catch((err) => {
-        if(err.response.status === 401 || 403)
+        if(err.response.status === 401 || err.response.status === 403)
         {
           const errData = err.response;
           const errorMsg = errData.status + ": " + errData.data.message;
           navigate(`/login/${errorMsg}`);
+        }
+        else if(err.response.status === 500)
+        {
+          console.log(err.response.data.message);
         }
       });
   }, []);
@@ -77,7 +81,8 @@ function MyTripsPage() {
       end: t.end_date,
       status: t.completed,
       isGroupTrip: false,
-      type: t.type
+      type: t.type,
+      userItineraryType: t.usertype
     }));
 
     setTrips((prev) => {
@@ -188,6 +193,7 @@ function MyTripsPage() {
   const deleteTripConfirmed = async () => {
     if (!tripToDelete) return;
 
+    // For group trips
     if (tripToDelete.isGroupTrip === true) {
       setTrips((prev) => prev.filter((t) => t.id !== tripToDelete.id));
       setShowDeleteConfirm(false);
@@ -198,7 +204,10 @@ function MyTripsPage() {
       return;
     }
 
-    try {
+    // If button is delete
+    if(tripToDelete.userItineraryType === "host")
+    {
+      try {
       const response = await axios.delete(
         "http://localhost:8080/Itinerary/DeleteItinerary",
         {
@@ -218,12 +227,42 @@ function MyTripsPage() {
         setErrorMsg("Delete failed.");
         setShowDeleteConfirm(false);
       }
-    } catch (err) {
-      console.error(err);
-      setErrorMsg("Delete failed (server error).");
-      setShowDeleteConfirm(false);
+      } catch (err) {
+        console.error(err);
+        setErrorMsg("Delete failed (server error).");
+        setShowDeleteConfirm(false);
+      }
+    }
+    // If button is exit
+    else if(tripToDelete.userItineraryType === "visitor"){
+      try {
+      const response = await axios.delete(
+        "http://localhost:8080/Itinerary/ExitItinerary",
+        {
+          data: { itineraryid: tripToDelete.id },
+          withCredentials: true
+        }
+      );
+
+      if (response.data === true) {
+        setTrips((prev) => prev.filter((t) => t.id !== tripToDelete.id));
+        setShowDeleteConfirm(false);
+        setTripToDelete(null);
+
+        setSuccessMsg("Trip deleted successfully!");
+        setTimeout(() => setSuccessMsg(""), 1500);
+      } else {
+        setErrorMsg("Delete failed.");
+        setShowDeleteConfirm(false);
+      }
+      } catch (err) {
+        console.error(err);
+        setErrorMsg("Delete failed (server error).");
+        setShowDeleteConfirm(false);
+      }
     }
   };
+
 
   //Apply filters
   const filteredTrips = trips.filter((t) => {
@@ -378,12 +417,13 @@ function MyTripsPage() {
                   View
                 </button>
 
-                <button
-                  className="delete-btn"
-                  onClick={() => requestDeleteTrip(trip.id)}
-                >
+                {trip.userItineraryType === "host" && <button className="delete-btn" onClick={() => requestDeleteTrip(trip.id)}>
                   Delete
-                </button>
+                </button>}
+
+                {trip.userItineraryType === "visitor" && <button className="delete-btn" onClick={() => requestDeleteTrip(trip.id)}>
+                  Exit
+                </button>}
               </div>
             </div>
           ))}
@@ -468,12 +508,12 @@ function MyTripsPage() {
         <div className="modal-overlay">
           <div className="modal-box small">
             <h2 className="modal-title">
-              {tripToDelete?.isGroupTrip ? "Leave Group Trip" : "Confirm Delete"}
+              {(tripToDelete?.isGroupTrip || tripToDelete.userItineraryType === "visitor") ? "Exit Group Trip" : "Confirm Delete"}
             </h2>
 
             <p>
-              {tripToDelete?.isGroupTrip
-                ? "Are you sure you want to leave this group trip?"
+              {(tripToDelete?.isGroupTrip || tripToDelete.userItineraryType === "visitor")
+                ? "Are you sure you want to exit this group trip?"
                 : "Are you sure you want to delete this trip?"}
             </p>
 
@@ -486,7 +526,7 @@ function MyTripsPage() {
               </button>
 
               <button className="modal-delete" onClick={deleteTripConfirmed}>
-                {tripToDelete?.isGroupTrip ? "Leave" : "Delete"}
+                {(tripToDelete?.isGroupTrip || tripToDelete.userItineraryType === "visitor") ? "Exit" : "Delete"}
               </button>
             </div>
           </div>
